# lib/api-zod/src/generated/types/appointmentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { AppointmentUpdateStatus } from './appointmentUpdateStatus';

export interface AppointmentUpdate {
  status?: AppointmentUpdateStatus;
  doctorName?: string;
  department?: string;
  date?: string;
  timeSlot?: string;
  notes?: string;
}