# lib/api-zod/src/generated/types/labReportInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { LabReportInputItemsItem } from './labReportInputItemsItem';

export interface LabReportInput {
  patientId: number;
  referredBy: string;
  reportDate: string;
  notes?: string;
  items: LabReportInputItemsItem[];
}