# lib/api-zod/src/generated/types/patient.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Hospital Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { PatientType } from './patientType';

export interface Patient {
  id: number;
  patientId: string;
  name: string;
  age: number;
  gender: string;
  phone: string;
  address: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  bloodGroup?: string | null;
  type: PatientType;
  /** @nullable */
  doctorName?: string | null;
  /** @nullable */
  department?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}